updated diagram_agent

+++ diagram_agent.py
@@ -2,8 +2,6 @@ import streamlit as st
 import subprocess
 import os
 from pathlib import Path
-import base64
-from data import sample_plantuml
 import openai
 
 # Path to the PlantUML .jar file
@@ -12,6 +10,10 @@ plantuml_jar_path = './plantuml.jar'
 # Connect to OpenAI key
 openai.api_key = st.secrets["OPENAI_API_KEY"]
 
+# Initialize session state for PlantUML code
+if 'plantuml_code' not in st.session_state:
+    st.session_state['plantuml_code'] = ""
+
 # Function to convert natural language instruction to PlantUML code using OpenAI
 def nl_to_plantuml(nl_instruction):
     try:
@@ -19,7 +21,7 @@ def nl_to_plantuml(nl_instruction):
         openai_response = openai.chat.completions.create(
             model="gpt-4-turbo-preview",
             messages=[
-                {"role": "system", "content": "You are a professional PlantUML coder. Use aws-orange theme. Output PlantUML code only and explain nothing."},
+                {"role": "system", "content": "You are a professional PlantUML coder. Include a title. Use aws-orange theme. Use note if needed to explain more details. You MUST Output PlantUML code only and explain nothing."},
                 {"role": "user", "content": nl_instruction}
             ],
             temperature=0.5,
@@ -31,7 +33,7 @@ def nl_to_plantuml(nl_instruction):
         st.error(f"An error occurred with the OpenAI API: {e}")
         return None
 
-#  Function to generate UML diagram from PlantUML code
+# Function to generate UML diagram from PlantUML code
 def generate_uml_diagram(plantuml_code):
     # Write the PlantUML code to a temporary file
     temp_file = 'temp.puml'
@@ -79,32 +81,36 @@ def get_image_download_link(img_path):
     return btn
 
 # Streamlit application layout
-st.title('PlantUML Diagram Generator')
+st.title('Diagram Generator')
 
 # Text area for user to enter natural language instructions
-nl_instruction = st.text_area(label="Describe your requirements in natural language:", height=300, placeholder="Help me create an instruciton on how to make a muffin")
+nl_instruction = st.text_area(label="Describe your requirements in natural language:", height=300, value="Help me create an instruction on how to make a muffin")
 
 # Button to convert natural language to PlantUML code
 if st.button("Convert to PlantUML"):
     plantuml_code = nl_to_plantuml(nl_instruction)
     if plantuml_code:
+        st.session_state['plantuml_code'] = plantuml_code
         st.success("Successfully converted to PlantUML code.")
         st.code(plantuml_code)
-else:
-    # Use sample PlantUML code if no conversion has been done yet
-    plantuml_code = sample_plantuml
+    else:
+        st.error("There was an error generating your PlantUML code")
 
 # Text area for user to enter or edit PlantUML code
-plantuml_code = st.text_area(label="Your PlantUML code:", value=plantuml_code, height=300)
+plantuml_code = st.text_area("Edit the PlantUML code here:", value=st.session_state['plantuml_code'], height=300)
+
+# Update session state when the user edits the code
+st.session_state['plantuml_code'] = plantuml_code
 
 # Automatically generate and display the diagram when the code changes
-output_file_path = generate_uml_diagram(plantuml_code)
-if output_file_path:
-    # Display the generated diagram
-    st.image(str(output_file_path), caption='Generated UML Diagram', use_column_width=True)
-    
-    # Provide a download button for the image
-    get_image_download_link(str(output_file_path))
-    
-    # Remove the output diagram file after displaying it
-    os.remove(output_file_path)
+if plantuml_code:
+    output_file_path = generate_uml_diagram(plantuml_code)
+    if output_file_path:
+        # Display the generated diagram
+        st.image(str(output_file_path), caption='Generated UML Diagram', use_column_width=False)
+        
+        # Provide a download button for the image
+        get_image_download_link(str(output_file_path))
+        
+        # Remove the output diagram file after displaying it
+        os.remove(output_file_path)
